refactor(agent): extract shared CLI utilities into cli-shared.ts — claude-cli.ts + codex-cli.ts
Removes 6 duplicated utility functions (createZeroUsage, normalizeUsage,
extractTextFromUnknown, stripToolCallJsonBlock, toUsageMetadata, TOOL_CALL_FORMAT)
from both CLI provider files. Unified normalizeUsage handles all field name
variants across providers (cache_read_input_tokens, cached_input_tokens, etc.).
